fix issue from not pulling before making changes
