+++ src/dask_capture.py
@@ -4,7 +4,7 @@ import warnings
 import pandas as pd
 import pickle
 
-from dask_md_objs import SchedulerEvent, TaskHandler, WXferEvent
+from dask_md_objs import SchedulerEvent, TaskHandler, WXferEvent, WorkerEvent
 
 ## TODO :: REMOVE ipykernel from UV
 
@@ -57,6 +57,21 @@ def extract_worker_xfer_metadata(worker_xfer_file: str, debug: bool = False, th:
 
     return th
 
+def extract_worker_metadata(worker_file: str, debug: bool=False, th:TaskHandler = None) -> TaskHandler :
+    worker_dat = pd.read_csv(worker_file)
+    nrow, ncol = worker_dat.shape
+
+    if th is None :
+        th = TaskHandler()
+    for i in range(0, nrow) :
+        th.add_event(WorkerEvent(worker_dat.iloc(axis=0)[i]))
+
+    if debug :
+        print(len(th.tasks))
+        print(th.tasks[list(th.tasks.keys())[0]])
+
+    return th
+
 if __name__ == "__main__":
     sched_file = file_meta_info["sched_file"][1]
     worker_file = file_meta_info["worker_file"][1]
@@ -90,6 +105,9 @@ if __name__ == "__main__":
     th = TaskHandler()
     extract_scheduler_metadata(sched_file, debug, th)
     extract_worker_xfer_metadata(worker_xfer_file, debug, th)
+    extract_worker_metadata(worker_trans_file, debug, th)
+
+    th.sort_tasks_by_time()
 
     if output_file is not None :
        with open(output_file, "w") as f:
